// src/lib/badges.generated.ts
export const BASELINE_GENERATED_AT = '2026-07-30'

export const BASELINE_BADGES: Record<string, BadgeMeta> = {
// ...
  bits: {
    label: 'Bits',
    uuid: '73b5c3fb-24f9-4a82-a852-2f475b59411c',
    perVersion: {
      '1': '73b5c3fb-24f9-4a82-a852-2f475b59411c',
      '100': '09d93036-e7ce-431c-9a9e-7044297133f2',
      '1000': '0d85a29e-79ad-4c63-a285-3acd2c66f2ba',
      '5000': '57cd97fc-3e9e-4c6d-9d41-60147137234e',
      '10000': '68af213b-a771-4124-b6e3-9bb6d98aa732',
      '25000': '64ca5920-c663-4bd8-bfb1-751b4caea2dd',
      '50000': '62310ba7-9916-4235-9eba-40110d67f85d',
      '75000': 'ce491fa4-b24f-4f3b-b6ff-44b080202792',
      '100000': '96f0540f-aa63-49e1-a8b3-259ece3bd098',
      '200000': '4a0b90c4-e4ef-407f-84fe-36b14aebdbb6',
      '300000': 'ac13372d-2e94-41d1-ae11-ecd677f69bb6',
      '400000': 'a8f393af-76e6-4aa2-9dd0-7dcc1c34f036',
      '500000': 'f6932b57-6a6e-4062-a770-dfbd9f4302e5',
      '600000': '4d908059-f91c-4aef-9acb-634434f4c32e',
      '700000': 'a1d2a824-f216-4b9f-9642-3de8ed370957',
      '800000': '5ec2ee3e-5633-4c2a-8e77-77473fe409e6',
      '900000': '088c58c6-7c38-45ba-8f73-63ef24189b84',
      '1000000': '494d1c8e-c3b2-4d88-8528-baff57c9bd3f',
      '1250000': 'ce217209-4615-4bf8-81e3-57d06b8b9dc7',
      '1500000': 'c4eba5b4-17a7-40a1-a668-bc1972c1e24d',
      '1750000': '183f1fd8-aaf4-450c-a413-e53f839f0f82',
      '2000000': '7ea89c53-1a3b-45f9-9223-d97ae19089f2',
      '2500000': 'cf061daf-d571-4811-bcc2-c55c8792bc8f',
      '3000000': '5671797f-5e9f-478c-a2b5-eb086c8928cf',
      '3500000': 'c3d218f5-1e45-419d-9c11-033a1ae54d3a',
      '4000000': '79fe642a-87f3-40b1-892e-a341747b6e08',
      '4500000': '736d4156-ac67-4256-a224-3e6e915436db',
      '5000000': '3f085f85-8d15-4a03-a829-17fca7bf1bc2',
    },
// ...
  },
// ...
}